removed garbage

+++ buildrun.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 
-#clear
-gcc -Wall  -c src/*.c && gcc -Wall  -o emulator *.o && rm *.o && ./emulator
+clear && gcc -Wall  -c src/*.c && gcc -Wall  -o emulator *.o && rm *.o && ./emulator
